+++ src/main.js
@@ -33,6 +33,7 @@ export { default as Input } from './components/Input';
 export { default as Label } from './components/Label';
 export { default as Link } from './components/Link';
 export { default as ListGroup } from './components/ListGroup';
+export { default as ManagerSideNav } from './components/ManagerSideNav';
 export { default as OverlayWrapper } from './components/OverlayWrapper';
 export { default as Pagination } from './components/Pagination';
 export { default as PaginationControls } from './components/PaginationControls';